Created BaseSpectrum class
